Bank Account + Wallet naming; unified money hub + inline borrow
- Rename: persistent money → "Bank Account" (💰); Cash Game run + challenge
  ante → "Wallet" (👛), replacing "Stack"/🪙. Mental model: Bank Account is
  savings; you move money into a Wallet to play.
- New shared LoanPanel.svelte: borrow/repay UI extracted from the bank page,
  reused by both /bank and the in-app money hub (single source of truth).
- Money hub: tapping your Bank Account (top money, menu or mid-game) opens one
  sheet with balance + Borrow/Repay + recent activity + "See full ledger" → /bank.
  Menu chip now opens the hub instead of routing away.
- Buy-in wall: short on a Cash Game buy-in → borrow exactly the shortfall
  (+25% fee), confirm with PIN, and the run auto-starts — no detour to /bank.
- Ledger labels: remap to live V2 reasons (cashgame_buyin/cashout, blitz_*,
  loan_*, etc.); drop the ugly auto-formatted fallbacks.

Spec: docs/superpowers/specs/2026-07-07-bank-account-loans-ui-design.md
